requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-socks==0.8.4
aiosignal==1.3.1
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
idna==3.7
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
matrix-nio==0.24.0
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
peewee==3.17.5
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.2
python-olm==3.2.16
python-socks==2.4.4
pytz==2024.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
rsa==4.9
soupsieve==2.5
tqdm==4.66.4
unpaddedbase64==2.1.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
